requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
click==8.0.4
colorama==0.4.4
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
mysqlclient==2.1.0
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.6
watchgod==0.7
websockets==10.2
